requirements.txt
asgiref==3.8.1
Django==4.2
django-filter==25.1
python-decouple==3.8
sqlparse==0.5.3
